+++ notebook/terminal/api_handlers.py
@@ -1,7 +1,8 @@
 import json
 from tornado import web, gen
 from ..base.handlers import APIHandler
-from ..utils import url_path_join
+from notebook.prometheus.metrics import TERMINAL_CURRENTLY_RUNNING_TOTAL
+
 
 class TerminalRootHandler(APIHandler):
     @web.authenticated
@@ -9,12 +10,16 @@ class TerminalRootHandler(APIHandler):
         tm = self.terminal_manager
         terms = [{'name': name} for name in tm.terminals]
         self.finish(json.dumps(terms))
+        TERMINAL_CURRENTLY_RUNNING_TOTAL.set(
+            len(terms)
+        )
 
     @web.authenticated
     def post(self):
         """POST /terminals creates a new terminal and redirects to it"""
         name, _ = self.terminal_manager.new_named_terminal()
         self.finish(json.dumps({'name': name}))
+        TERMINAL_CURRENTLY_RUNNING_TOTAL.inc()
 
 
 class TerminalHandler(APIHandler):
@@ -36,5 +41,6 @@ class TerminalHandler(APIHandler):
             yield tm.terminate(name, force=True)
             self.set_status(204)
             self.finish()
+            TERMINAL_CURRENTLY_RUNNING_TOTAL.dec(1)
         else:
             raise web.HTTPError(404, "Terminal not found: %r" % name)
